Actually move back all queries back to original places of implementation
